+++ src/components/projects/Projects.styled.js
@@ -1,10 +1,6 @@
 import styled from "styled-components";
 
 export const StyledProjectsSection = styled.section`
-  // display: flex;
-  // flex-direction: column;
-  // align-items: center;
-
   .bigButton {
     margin: 10em;
     ${({ theme }) => theme.bigButton};
