+++ src/components/socials/Socials.jsx
@@ -3,26 +3,11 @@ import React from "react";
 import { StyledSocialList } from "./Socials.styled";
 import { Side } from "../";
 import { Icon } from "../icons";
+import { socialMedia } from "../../config";
 
 const Socials = () => {
-  let socialMedia = [
-    {
-      name: "GitHub",
-      url: "https://github.com/fevazquez",
-    },
-    {
-      name: "LinkedIn",
-      url: "https://www.linkedin.com/in/",
-    },
-    {
-      name: "Instagram",
-      url: "https://www.instagram.com/itsfernanflow/",
-    },
-  ];
-
-  const isHome = true;
   return (
-    <Side isHome={isHome} orientation="left">
+    <Side orientation="left">
       <StyledSocialList>
         {socialMedia &&
           socialMedia.map(({ url, name }, i) => (
